init support for taking determinant.

+++ src/geometry/point.rs
@@ -1,13 +1,13 @@
 
 #[derive(Debug, Eq, Ord, PartialEq, PartialOrd,Clone)]
 pub struct Point {
-    pub x: i64,
-    pub y: i64
+    pub x: f64,
+    pub y: f64
 }
 
 
 impl Point {
-    pub fn new(x:i64, y: i64) -> Self {
+    pub fn new(x:f64, y: f64) -> Self {
         Point {
             x,
             y
